moved geopedia functionalities into a subfolder
