+++ docs/source/geopedia/core.rst
@@ -1,0 +1,7 @@
+core
+====
+
+.. automodule:: sentinelhub.geopedia.core
+    :members:
+    :inherited-members:
+    :show-inheritance:
