+++ docs/source/geopedia/request.rst
@@ -1,0 +1,7 @@
+request
+=======
+
+.. automodule:: sentinelhub.geopedia.request
+    :members:
+    :inherited-members:
+    :show-inheritance:
